working bugs
