+++ honeybee_energy/construction/shade.py
@@ -124,7 +124,8 @@ class ShadeConstruction(object):
         dictionary should be of a standard Python type to ensure correct
         serialization of the object to/from JSON (eg. str, float, int, list, dict)
         """
-        return self._user_data
+        if self._user_data is not None:
+            return self._user_data
 
     @user_data.setter
     def user_data(self, value):
@@ -173,7 +174,7 @@ class ShadeConstruction(object):
         if 'display_name' in data and data['display_name'] is not None:
             new_obj.display_name = data['display_name']
         if 'user_data' in data and data['user_data'] is not None:
-            new_obj._user_data = data['user_data']
+            new_obj.user_data = data['user_data']
         return new_obj
 
     def to_idf(self, host_shade_identifier):
@@ -217,7 +218,7 @@ class ShadeConstruction(object):
         if self._display_name is not None:
             base['display_name'] = self.display_name
         if self._user_data is not None:
-            base['user_data'] = self._user_data
+            base['user_data'] = self.user_data
         return base
 
     def duplicate(self):
@@ -243,7 +244,7 @@ class ShadeConstruction(object):
             self.identifier, self._solar_reflectance, self._visible_reflectance,
             self._is_specular)
         new_con._display_name = self._display_name
-        new_con.user_data = None if self._user_data is None else self._user_data.copy()
+        new_con._user_data = None if self._user_data is None else self._user_data.copy()
         return new_con
 
     def __key(self):
